test(e2e): migrate zone egress universal external services
